fix(pg): INTEGER→BIGINT drift on meta_optimizer_snapshots.runs_included
The Rust struct (meta_optimizer/snapshots.rs:24) declares
`runs_included: i64`, and every call site reads via `r.get::<_, i64>(7)`.
The live column was INTEGER (int4), so reads panicked with
"error retrieving column 7: error deserializing column 7" the first
time get_latest_baseline_snapshot or any of its 4 sibling readers
returned a row — reproducing across instances on 2026-04-21 and
2026-04-22 (3 captured panics in runner-last-panic.txt).

Root cause mirrors the schema-drift-audit pattern: the column was
originally declared INTEGER in both schema.pg.sql and
ensure_tables, so CREATE TABLE IF NOT EXISTS never updates it. Fix
is the same as v9's BIGINT repairs:

- New migration v26: ALTER COLUMN TYPE BIGINT USING ::bigint (lossless)
- schema.pg.sql canonical: INTEGER → BIGINT
- ensure_tables CREATE TABLE: INTEGER → BIGINT (for fresh installs)

Verified clean inside a docker-exec transaction before committing.

+++ src-tauri/schema.pg.sql
@@ -2260,7 +2260,7 @@ CREATE TABLE IF NOT EXISTS meta_optimizer_snapshots (
     metrics_json TEXT NOT NULL,
     breakdown_json TEXT DEFAULT '{}',
     recommendation_id TEXT,
-    runs_included INTEGER NOT NULL DEFAULT 0,
+    runs_included BIGINT NOT NULL DEFAULT 0,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
 CREATE INDEX IF NOT EXISTS idx_meta_optimizer_snapshots_type ON meta_optimizer_snapshots(snapshot_type);
